Remove unnecessary files and refactor code structure
