--- .env.example
@@ -1,2 +1,0 @@
-SR_TOKEN=token
-REDIS_URL=redis://localhost:6379
