--- .prettierrc
@@ -1,25 +1,0 @@
-{
-  "singleQuote": true,
-  "printWidth": 100,
-  "overrides": [
-    {
-      "files": "*.md(x)?",
-      "options": {
-        "printWidth": 80,
-        "proseWrap": "always"
-      }
-    }
-  ],
-  "plugins": ["prettier-plugin-packagejson", "@ianvs/prettier-plugin-sort-imports"],
-  "importOrder": [
-    "^hono",
-    "^@hono",
-    "",
-    "<BUILTIN_MODULES>",
-    "<THIRD_PARTY_MODULES>",
-    "",
-    "^@(serverless-redis)/(.*)$",
-    "",
-    "^[./]"
-  ]
-}
