+++ README.md
@@ -1,13 +1,21 @@
 # Serverless Redis with Upstash compatibility
 
-This repo provides an HTTP server that interacts with the Redis database via HTTP requests. It supports compatibility with Upstash, a cloud service that provides Redis.
+This repo provides an HTTP server that interacts with the Redis database via
+HTTP requests. It supports compatibility with Upstash, a cloud service that
+provides Redis.
 
 [![Deploy on Railway](https://railway.app/button.svg)](https://railway.app/template/hBFwO4?referralCode=73cYCO)
 
 ### Technologies used:
-- **Hono**: A lightweight framework for building high-performance HTTP servers in TypeScript. In this template, Hono is used to handle HTTP requests.
-- **Nx**: A project management tool that helps organize code and dependencies. It is used to structure the application and run tasks such as build, testing, and deployment.
+
+- **Hono**: A lightweight framework for building high-performance HTTP servers
+  in TypeScript. In this template, Hono is used to handle HTTP requests.
+- **Nx**: A project management tool that helps organize code and dependencies.
+  It is used to structure the application and run tasks such as build, testing,
+  and deployment.
 
 ### Why use this template:
+
 - Simple setup and high performance thanks to Hono.
-- Nx helps manage the project and dependencies, making development more convenient.
+- Nx helps manage the project and dependencies, making development more
+  convenient.
