--- apps/api/jest.config.ts
@@ -1,12 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'api',
-  preset: '../../jest.preset.js',
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]s$': '@swc/jest',
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/apps/api',
-  passWithNoTests: true,
-};
