--- apps/api/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "@serverless-resdis/api-source",
-  "version": "1.2.0",
-  "private": true
-}
