--- package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@serverless-redis/source",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "scripts": {},
-  "dependencies": {
-    "@hono/node-server": "^1.12.0",
-    "hono": "^4.4.11",
-    "ioredis": "^5.4.1",
-    "tslib": "^2.3.0"
-  },
-  "devDependencies": {
-    "@ianvs/prettier-plugin-sort-imports": "^4.3.0",
-    "@nx/esbuild": "19.4.0",
-    "@nx/eslint": "19.4.0",
-    "@nx/eslint-plugin": "19.4.0",
-    "@nx/jest": "19.4.0",
-    "@nx/js": "19.4.0",
-    "@nx/node": "^19.4.0",
-    "@nx/vite": "19.4.0",
-    "@nx/web": "19.4.0",
-    "@nx/workspace": "19.4.0",
-    "@swc-node/register": "~1.9.1",
-    "@swc/core": "~1.5.7",
-    "@swc/helpers": "~0.5.11",
-    "@swc/jest": "~0.2.36",
-    "@types/jest": "^29.4.0",
-    "@types/node": "18.16.9",
-    "@typescript-eslint/eslint-plugin": "^7.3.0",
-    "@typescript-eslint/parser": "^7.3.0",
-    "@upstash/redis": "^1.32.0",
-    "@vitest/coverage-v8": "^1.0.4",
-    "@vitest/ui": "^1.3.1",
-    "autoprefixer": "10.4.13",
-    "babel-jest": "^29.4.1",
-    "esbuild": "^0.19.2",
-    "eslint": "~8.57.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "2.27.5",
-    "jest": "^29.4.1",
-    "jest-environment-jsdom": "^29.4.1",
-    "jest-environment-node": "^29.4.1",
-    "nx": "19.4.0",
-    "postcss": "8.4.38",
-    "prettier": "^3.3.2",
-    "prettier-plugin-packagejson": "^2.5.0",
-    "tailwindcss": "3.4.3",
-    "ts-jest": "^29.1.0",
-    "ts-node": "10.9.1",
-    "typescript": "~5.4.2",
-    "vite": "^5.0.0",
-    "vitest": "^1.3.1"
-  }
-}
