+++ src/lib/utils/is-encoding-enabled.ts
@@ -1,4 +1,4 @@
-import { HonoRequest } from 'hono';
+import type { HonoRequest } from 'npm:hono';
 
 export function isEncodingEnabled(req: HonoRequest) {
   const encodingHeader = req.header('upstash-encoding');
