+++ src/lib/utils/is-result-array.ts
@@ -1,4 +1,4 @@
-import { isObject } from './is-object';
+import { isObject } from './is-object.ts';
 
 export function isResultArray(value: unknown) {
   return Array.isArray(value) && value.length && value.every(isObject);
