+++ src/main.test.ts
@@ -1,0 +1,5 @@
+import { assertEquals } from '@std/assert';
+
+Deno.test(function addTest() {
+  assertEquals(5, 5);
+});
