--- vitest.workspace.ts
@@ -1,1 +1,0 @@
-export default ['**/*/vite.config.ts', '**/*/vitest.config.ts'];
